added validations for enterprise features in ce
