+++ internal/multicluster/internal/types/validation_exported_services_ee.go
@@ -13,6 +13,35 @@ import (
 	"github.com/hashicorp/go-multierror"
 )
 
+func ValidateExportedServices(res *pbresource.Resource) error {
+	var exportedService pbmulticluster.ExportedServices
+
+	if err := res.Data.UnmarshalTo(&exportedService); err != nil {
+		return resource.NewErrDataParse(&exportedService, err)
+	}
+
+	var merr error
+
+	if exportedService.Services == nil || len(exportedService.Services) == 0 {
+		merr = multierror.Append(merr, resource.ErrInvalidField{
+			Name:    "services",
+			Wrapped: fmt.Errorf("at least one service must be set"),
+		})
+	}
+
+	return merr
+}
+
+func ValidateNamespaceExportedServices(res *pbresource.Resource) error {
+	// No-op
+	return nil
+}
+
+func ValidatePartitionExportedServices(res *pbresource.Resource) error {
+	// No-op
+	return nil
+}
+
 func ValidateComputedExportedServices(res *pbresource.Resource) error {
 	var computedExportedServices pbmulticluster.ComputedExportedServices
